feat: refactor CandidateNew to CandidateUpsert
