+++ app/javascript/src/views/index.js
@@ -1,4 +1,4 @@
 export { default as Candidates } from './Candidates.jsx'
 export { default as Candidate } from './Candidate.jsx'
-export { default as CandidateNew } from './CandidateNew.jsx'
+export { default as CandidateUpsert } from './CandidateUpsert.jsx'
 export { default as Home } from './Home.jsx'
